Support Animated Gifs Inline
This implements a new Paintable I borrowed from a forum and converted
into Vala. The PixbufAnimation, if given a non-animated image simply
returns it as a one-frame, non-looping, animation and so it shouldn't be
a lot more overhead than the original method. It just draws once and
then never updates again.

So it appears to work with existing images, but also handles animation!

Closes #386